refactor($CarName, RacingNumber): create `vo` package and move
